requirements.txt
streamlit
requests
folium
streamlit-folium
geopandas
pandas
openpyxl
st-annotated-text
matplotlib

